Extract schedule ascii code
The ScheduleView class was really unmanageable in size due to the ASCII
layouting code, and there is no need to keep that in the view class.
